Add transform and replace; redefine map using transform
Transform is a more generic variant of map that can operate on a
stream's values or state to yield a new stream. Map may be defined in
terms of transform.

The replace function is like map, only it takes a new state processing
function and some function to apply to a stream's state to yield a new
stream.

Currently, there's a bit of an oddity in the transform implementation
whereby we need to copy a function ref--the generated c winds up using
memory incorrectly otherwise.

I also added a test for `replace`